+++ packages/components/src/components/Navigation/Navigation.tsx
@@ -1,19 +1,25 @@
-import type { ComponentProps, FC, PropsWithChildren } from "react";
+import type { ComponentProps, PropsWithChildren } from "react";
 import React from "react";
 import styles from "./Navigation.module.scss";
 import clsx from "clsx";
-import type { PropsContext } from "@/lib/propsContext";
+import { type PropsContext } from "@/lib/propsContext";
 import { PropsContextProvider } from "@/lib/propsContext";
 import type { PropsWithClassName } from "@/lib/types/props";
-import { TunnelExit } from "@mittwald/react-tunnel";
+import { TunnelExit, TunnelProvider } from "@mittwald/react-tunnel";
+import {
+  flowComponent,
+  type FlowComponentProps,
+} from "@/lib/componentFactory/flowComponent";
+import ClearPropsContext from "@/components/ClearPropsContext";
 
 export interface NavigationProps
   extends PropsWithChildren<ComponentProps<"nav">>,
-    PropsWithClassName {}
+    PropsWithClassName,
+    FlowComponentProps<HTMLElement> {}
 
 /** @flr-generate all */
-export const Navigation: FC<NavigationProps> = (props) => {
-  const { className, children, ...rest } = props;
+export const Navigation = flowComponent("Navigation", (props) => {
+  const { className, children, ref, ...rest } = props;
 
   const rootClassName = clsx(styles.navigation, className);
 
@@ -33,15 +39,19 @@ export const Navigation: FC<NavigationProps> = (props) => {
   };
 
   return (
-    <nav className={rootClassName} role="navigation" {...rest}>
-      <PropsContextProvider props={propsContext} mergeInParentContext>
-        <ul>
-          <TunnelExit id="links" />
-        </ul>
-        {children}
-      </PropsContextProvider>
-    </nav>
+    <ClearPropsContext>
+      <TunnelProvider>
+        <nav className={rootClassName} role="navigation" {...rest} ref={ref}>
+          <PropsContextProvider props={propsContext} mergeInParentContext>
+            <ul>
+              <TunnelExit id="links" />
+            </ul>
+            {children}
+          </PropsContextProvider>
+        </nav>
+      </TunnelProvider>
+    </ClearPropsContext>
   );
-};
+});
 
 export default Navigation;